style: Aligned "Show full agenda" button to the left, fixes #65  💄
Fixes #65

Kaan said that the button was better noticable on the left during the usertest.
I changed the button to the left by deleting the margin-left: auto, and added padding on the top so the button didn't stick to the text above. I also deleted the margin-left in the label tag, because in this way the label and span were aligned together nicely.