style: more clearer & consistent states for interactive elements, fixes #63 💄
fixes #63

I first fixed issue #69 , so i could change the states for the nav button. Nav button has the same states as the other buttons on the page. Further i gave interactive text elements a underline when hovering, giving the user a indicator that its a clickable element. Also deleted the automatic focus-visible outline on the buttons, because it was not needed and ugly. Also moved the variable font settings to globalcss file.