fix: delete the hiding of scrollbar on windows OS, fixes #67 🐛
Fixes #67

Deleted the webkit-scrollbar (display: none) in the global.css. This was a bad decision as its bad for the accessibility. The scrollbar was hidden, because of this it doesn't show on the windows OS, while normally people see a scrollbar on their screen. It would confuse users, that normally use the scrollbar.