style: tweaked consistency in style #63 💄
Update more style consistency, i saw somethings that could improve. The <a>'s in molecules/Semester & organisms/Nav focus(-visible) & hover weren't good visible with only a colored underline, so i've made the next another color on the hover and focus as well.

In the organisms/Nav i've deleted the font-weight normal, because there is also a strong that is bold in the rest of the page. This fixes the issue of inconsistency.