feat: svg component class prop, fixes #78 ✨
fixes #78

Made a Prop with a class in the SVG of the component,  because of this there weren't 2 SVG's needed.

Both were Arrows, 1 to the left and 1 to the right. I made a prop which gives 1 the class right-arrow with transformX -1. This makes the left arrow turn to the right.